Automatically release superlocks when we no longer need them
We do this by making $Self->{superlock} a weakref to the lock
and returning it to the upper scope in begin_superlock().

As long as the upper scope hangs onto the lock scalar, they
have the lock.

This way, we don't have to make sure to always call end_superlock()
(which we weren't doing in cases like invalidArguments errors in
api_email_import() for example).